requirements.txt
django
djangorestframework
psycopg2-binary
pandas
openpyxl
celery
redis
